+++ src/registrar/templates/includes/member_domains_edit_table.html
@@ -21,8 +21,8 @@
 {% endif %}
 
 {% comment %} Stores the json endpoint in a url for easier access {% endcomment %}
-{% url 'get_member_domains_edit_json' as url %}
-<span id="get_member_domains_edit_json_url" class="display-none">{{url}}</span>
+{% url 'get_member_domains_json' as url %}
+<span id="get_member_domains_json_url" class="display-none">{{url}}</span>
 <section class="section-outlined member-domains margin-top-0 section-outlined--border-base-light" id="edit-member-domains">
     
   <h2>
