+++ src/registrar/views/member_domains_json.py
@@ -1,4 +1,5 @@
 import logging
+from django.db import models
 from django.http import JsonResponse
 from django.core.paginator import Paginator
 from django.shortcuts import get_object_or_404
@@ -28,11 +29,12 @@ class PortfolioMemberDomainsJson(PortfolioMemberDomainsPermission, View):
         objects = self.apply_search(objects, request)
         objects = self.apply_sorting(objects, request)
 
-        paginator = Paginator(objects, 10)
+        paginator = Paginator(objects, self.get_page_size(request))
         page_number = request.GET.get("page")
         page_obj = paginator.get_page(page_number)
 
-        domains = [self.serialize_domain(domain, request.user) for domain in page_obj.object_list]
+        member_id = request.GET.get("member_id")
+        domains = [self.serialize_domain(domain, member_id, request.user) for domain in page_obj.object_list]
 
         return JsonResponse(
             {
@@ -46,6 +48,23 @@ class PortfolioMemberDomainsJson(PortfolioMemberDomainsPermission, View):
             }
         )
 
+    def get_page_size(self, request):
+        """Gets the page size.
+        
+        If member_only, need to return the entire result set every time, so need
+        to set to a very large page size. If not member_only, this can be adjusted
+        to provide a smaller page size"""
+
+        member_only = request.GET.get("member_only", "false").lower() in ["true", "1"]
+        if member_only:
+            # This number needs to remain very high as the entire result set
+            # must be returned when member_only
+            return 1000
+        else:
+            # This number can be adjusted if we want to add pagination to the result page
+            # later
+            return 1000
+        
     def get_domain_ids_from_request(self, request):
         """Get domain ids from request.
 
@@ -86,13 +105,42 @@ class PortfolioMemberDomainsJson(PortfolioMemberDomainsPermission, View):
         return queryset
 
     def apply_sorting(self, queryset, request):
+        # Get the sorting parameters from the request
         sort_by = request.GET.get("sort_by", "name")
         order = request.GET.get("order", "asc")
-        if order == "desc":
-            sort_by = f"-{sort_by}"
-        return queryset.order_by(sort_by)
+        # Sort by 'checked' if specified, otherwise by the given field
+        if sort_by == "checked":
+            # Get list of checked ids from the request
+            checked_ids = request.GET.get("checkedDomainIds")
+            if checked_ids:
+                # Split the comma-separated string into a list of integers
+                checked_ids = [int(id.strip()) for id in checked_ids.split(",") if id.strip().isdigit()]
+            else:
+                # If no value is passed, set checked_ids to an empty list
+                checked_ids = []
+            # Annotate each object with a 'checked' value based on whether its ID is in checkedIds
+            queryset = queryset.annotate(
+                checked=models.Case(
+                    models.When(id__in=checked_ids, then=models.Value(True)),
+                    default=models.Value(False),
+                    output_field=models.BooleanField(),
+                )
+            )
+            # Add ordering logic for 'checked'
+            if order == "desc":
+                queryset = queryset.order_by("-checked","name")
+            else:
+                queryset = queryset.order_by("checked","name")
+        else:
+            # Handle other fields as normal
+            if order == "desc":
+                sort_by = f"-{sort_by}"
+            queryset = queryset.order_by(sort_by)
+        
+        return queryset
 
-    def serialize_domain(self, domain, user):
+
+    def serialize_domain(self, domain, member_id, user):
         suborganization_name = None
         try:
             domain_info = domain.domain_info
@@ -107,9 +155,18 @@ class PortfolioMemberDomainsJson(PortfolioMemberDomainsPermission, View):
         # Check if there is a UserDomainRole for this domain and user
         user_domain_role_exists = UserDomainRole.objects.filter(domain_id=domain.id, user=user).exists()
         view_only = not user_domain_role_exists or domain.state in [Domain.State.DELETED, Domain.State.ON_HOLD]
+
+        # Check if the specified member is the only member assigned as manager of domain
+        only_member_assigned_to_domain = False
+        if member_id:
+            member_domain_role_count = UserDomainRole.objects.filter(domain_id=domain.id).count()
+            member_domain_role_exists = UserDomainRole.objects.filter(domain_id=domain.id, user_id=member_id).exists()
+            only_member_assigned_to_domain = member_domain_role_exists and member_domain_role_count == 1
+
         return {
             "id": domain.id,
             "name": domain.name,
+            "member_is_only_manager": only_member_assigned_to_domain,
             "expiration_date": domain.expiration_date,
             "state": domain.state,
             "state_display": domain.state_display(),
